[Block Library - Query Loop]: Reorganise inspector controls
